moave session-related processing
