--- neko/app/helpers/sessions_helper.rb
@@ -1,22 +1,0 @@
-module SessionsHelper
-  def log_in(user)
-    session[:user_id] = user.id
-  end
-
-  def logged_in?
-    current_user
-  end
-
-  def current_user
-    @current_user ||= User.find_by(id: session[:user_id]) if session[:user_id]
-  end
-
-  def current_user?(user)
-    user == current_user
-  end
-
-  def log_out
-    session.delete(:user_id)
-    @current_user = nil
-  end
-end
